🎨 [Structure] : Rename nestjs convention
