+++ src/members/dto/queryString.dto.ts
