+++ src/members/dto/update-member.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateMemberDto } from './create.member';
+import { CreateMemberDto } from './create-member.dto';
 
 export class UpdateMemberDto extends PartialType(CreateMemberDto) {}
