+++ src/members/members.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/common/prismaClient';
-import { WhereCondition } from './dto/whereCondition';
+import { WhereCondition } from './dto/whereCondition.dto';
 import { Prisma } from '@prisma/client';
 @Injectable()
 export class MembersRepository {
